test: parallelize chat API e2e tests
- Split .each() tests and individual tests into separate files
- Configure Vitest to run up to 10 tests in parallel
- Add concurrent mode for .each() test suites
- Update CLAUDE.md with new test structure documentation
