+++ apps/gateway/src/api-individual.e2e.ts
@@ -1,0 +1,649 @@
+import "dotenv/config";
+import { beforeEach, describe, expect, test } from "vitest";
+
+import { db, tables, eq } from "@llmgateway/db";
+import { models, providers } from "@llmgateway/models";
+
+import { app } from ".";
+import {
+	clearCache,
+	waitForLogs,
+	waitForLogByRequestId,
+	getProviderEnvVar,
+} from "./test-utils/test-helpers";
+
+// Helper function to generate unique request IDs for tests
+function generateTestRequestId(): string {
+	return `test-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
+}
+
+const logMode = process.env.LOG_MODE;
+
+describe("e2e individual tests", () => {
+	beforeEach(async () => {
+		await clearCache();
+
+		await Promise.all([
+			db.delete(tables.log),
+			db.delete(tables.apiKey),
+			db.delete(tables.providerKey),
+		]);
+
+		await Promise.all([
+			db.delete(tables.userOrganization),
+			db.delete(tables.project),
+		]);
+
+		await Promise.all([
+			db.delete(tables.organization),
+			db.delete(tables.user),
+			db.delete(tables.account),
+			db.delete(tables.session),
+			db.delete(tables.verification),
+		]);
+
+		await db.insert(tables.user).values({
+			id: "user-id",
+			name: "user",
+			email: "user",
+		});
+
+		await db.insert(tables.organization).values({
+			id: "org-id",
+			name: "Test Organization",
+			plan: "pro",
+		});
+
+		await db.insert(tables.userOrganization).values({
+			id: "user-org-id",
+			userId: "user-id",
+			organizationId: "org-id",
+		});
+
+		await db.insert(tables.project).values({
+			id: "project-id",
+			name: "Test Project",
+			organizationId: "org-id",
+			mode: "api-keys",
+		});
+
+		await db.insert(tables.apiKey).values({
+			id: "token-id",
+			token: "real-token",
+			projectId: "project-id",
+			description: "Test API Key",
+		});
+
+		for (const provider of providers) {
+			const envVarName = getProviderEnvVar(provider.id);
+			const envVarValue = envVarName ? process.env[envVarName] : undefined;
+			if (envVarValue) {
+				await createProviderKey(provider.id, envVarValue, "api-keys");
+				await createProviderKey(provider.id, envVarValue, "credits");
+			}
+		}
+	});
+
+	async function createProviderKey(
+		provider: string,
+		token: string,
+		keyType: "api-keys" | "credits" = "api-keys",
+	) {
+		const keyId =
+			keyType === "credits" ? `env-${provider}` : `provider-key-${provider}`;
+		await db.insert(tables.providerKey).values({
+			id: keyId,
+			token,
+			provider: provider.replace("env-", ""), // Remove env- prefix for the provider field
+			organizationId: "org-id",
+		});
+	}
+
+	async function validateLogByRequestId(requestId: string) {
+		const log = await waitForLogByRequestId(requestId);
+
+		if (logMode) {
+			console.log("log", JSON.stringify(log, null, 2));
+		}
+
+		expect(log.usedProvider).toBeTruthy();
+		expect(log.errorDetails).toBeNull();
+		expect(log.finishReason).not.toBeNull();
+		expect(log.unifiedFinishReason).not.toBeNull();
+		expect(log.unifiedFinishReason).toBeTruthy();
+		expect(log.usedModel).toBeTruthy();
+		expect(log.requestedModel).toBeTruthy();
+
+		return log;
+	}
+
+	function validateResponse(json: any) {
+		expect(json).toHaveProperty("choices.[0].message.content");
+
+		expect(json).toHaveProperty("usage.prompt_tokens");
+		expect(json).toHaveProperty("usage.completion_tokens");
+		expect(json).toHaveProperty("usage.total_tokens");
+	}
+
+	test("JSON output mode error for unsupported model", async () => {
+		const envVarName = getProviderEnvVar("anthropic");
+		const envVarValue = envVarName ? process.env[envVarName] : undefined;
+		if (!envVarValue) {
+			console.log(
+				"Skipping JSON output error test - no Anthropic API key provided",
+			);
+			return;
+		}
+
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: "anthropic/claude-3-5-sonnet-20241022",
+				messages: [
+					{
+						role: "user",
+						content: "Hello",
+					},
+				],
+				response_format: { type: "json_object" },
+			}),
+		});
+
+		expect(res.status).toBe(400);
+
+		const text = await res.text();
+		expect(text).toContain("does not support JSON output mode");
+
+		await clearCache(); // Process logs BEFORE deleting data
+		await db.delete(tables.apiKey);
+		await db.delete(tables.providerKey);
+	});
+
+	test("JSON output mode error when 'json' not mentioned in messages", async () => {
+		const envVarName = getProviderEnvVar("openai");
+		const envVarValue = envVarName ? process.env[envVarName] : undefined;
+		if (!envVarValue) {
+			console.log(
+				"Skipping JSON output client error test - no OpenAI API key provided",
+			);
+			return;
+		}
+
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: "gpt-4o-mini",
+				messages: [
+					{
+						role: "user",
+						content: "Hello, give me a greeting response",
+					},
+				],
+				response_format: { type: "json_object" },
+			}),
+		});
+
+		expect(res.status).toBe(400);
+
+		const json = await res.json();
+		expect(json).toHaveProperty("error");
+		expect(json.error).toHaveProperty("message");
+		expect(json.error.message).toContain("'messages' must contain");
+		expect(json.error.message).toContain("the word 'json'");
+		expect(json.error).toHaveProperty("type", "invalid_request_error");
+
+		// Wait for logs to be processed
+		const logs = await waitForLogs(1);
+		expect(logs.length).toBe(1);
+
+		const log = logs[0];
+		expect(log.unifiedFinishReason).toBe("client_error");
+		expect(log.errorDetails).not.toBeNull();
+		expect(log.errorDetails).toHaveProperty("message");
+		expect((log.errorDetails as { message?: string })?.message).toContain(
+			"'messages' must contain",
+		);
+		expect((log.errorDetails as { message?: string })?.message).toContain(
+			"the word 'json'",
+		);
+	});
+
+	test("completions with llmgateway/auto in credits mode", async () => {
+		// require all provider keys to be set
+		for (const provider of providers) {
+			const envVarName = getProviderEnvVar(provider.id);
+			const envVarValue = envVarName ? process.env[envVarName] : undefined;
+			if (!envVarValue) {
+				console.log(
+					`Skipping llmgateway/auto in credits mode test - no API key provided for ${provider.id}`,
+				);
+				return;
+			}
+		}
+
+		await db
+			.update(tables.organization)
+			.set({ credits: "1000" })
+			.where(eq(tables.organization.id, "org-id"));
+
+		await db
+			.update(tables.project)
+			.set({ mode: "credits" })
+			.where(eq(tables.project.id, "project-id"));
+
+		await db.insert(tables.apiKey).values({
+			id: "token-credits",
+			token: "credits-token",
+			projectId: "project-id",
+			description: "Test API Key for Credits",
+		});
+
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer credits-token`,
+			},
+			body: JSON.stringify({
+				model: "llmgateway/auto",
+				messages: [
+					{
+						role: "user",
+						content: "Hello with llmgateway/auto in credits mode!",
+					},
+				],
+			}),
+		});
+
+		const json = await res.json();
+		console.log("response:", JSON.stringify(json, null, 2));
+		expect(res.status).toBe(200);
+		expect(json).toHaveProperty("choices.[0].message.content");
+
+		const logs = await waitForLogs(1);
+		expect(logs.length).toBe(1);
+		expect(logs[0].requestedModel).toBe("auto");
+		expect(logs[0].usedProvider).toBeTruthy();
+		expect(logs[0].usedModel).toBeTruthy();
+	});
+
+	test("completions with bare 'auto' model and credits", async () => {
+		await db
+			.update(tables.organization)
+			.set({ credits: "1000" })
+			.where(eq(tables.organization.id, "org-id"));
+
+		await db
+			.update(tables.project)
+			.set({ mode: "credits" })
+			.where(eq(tables.project.id, "project-id"));
+
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: "auto",
+				messages: [
+					{
+						role: "user",
+						content: "Hello! This is an auto test.",
+					},
+				],
+			}),
+		});
+
+		expect(res.status).toBe(200);
+		const json = await res.json();
+		expect(json).toHaveProperty("choices.[0].message.content");
+
+		const logs = await waitForLogs(1);
+		expect(logs.length).toBe(1);
+		expect(logs[0].requestedModel).toBe("auto");
+		expect(logs[0].usedProvider).toBeTruthy();
+		expect(logs[0].usedModel).toBeTruthy();
+	});
+
+	test.skip("/v1/chat/completions with bare 'custom' model", async () => {
+		const envVarName = getProviderEnvVar("openai");
+		const envVarValue = envVarName ? process.env[envVarName] : undefined;
+		if (!envVarValue) {
+			console.log("Skipping custom model test - no OpenAI API key provided");
+			return;
+		}
+
+		await db
+			.update(tables.organization)
+			.set({ credits: "1000" })
+			.where(eq(tables.organization.id, "org-id"));
+
+		await db
+			.update(tables.project)
+			.set({ mode: "credits" })
+			.where(eq(tables.project.id, "project-id"));
+
+		await db.insert(tables.providerKey).values({
+			id: "provider-key-custom",
+			provider: "llmgateway",
+			token: envVarValue,
+			baseUrl: "https://api.openai.com", // Use real OpenAI endpoint for testing
+			status: "active",
+			organizationId: "org-id",
+		});
+
+		await db.insert(tables.apiKey).values({
+			id: "token-id-2",
+			token: "real-token-2",
+			projectId: "project-id",
+			description: "Test API Key",
+		});
+
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer real-token-2`,
+			},
+			body: JSON.stringify({
+				model: "custom",
+				messages: [
+					{
+						role: "user",
+						content: "Hello! This is a custom test.",
+					},
+				],
+			}),
+		});
+
+		expect(res.status).toBe(200);
+
+		const json = await res.json();
+		expect(json).toHaveProperty("choices.[0].message.content");
+
+		const logs = await waitForLogs(1);
+		expect(logs.length).toBe(1);
+		expect(logs[0].requestedModel).toBe("custom");
+		expect(logs[0].usedProvider).toBe("llmgateway");
+		expect(logs[0].usedModel).toBe("custom");
+	});
+
+	test("Prompt tokens are never zero even when provider returns 0", async () => {
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: "openai/gpt-4o-mini",
+				messages: [
+					{
+						role: "user",
+						content: "ZERO_TOKENS test message",
+					},
+				],
+			}),
+		});
+
+		expect(res.status).toBe(200);
+
+		const json = await res.json();
+
+		// Verify we have usage information
+		expect(json).toHaveProperty("usage");
+		expect(json.usage).toHaveProperty("prompt_tokens");
+		expect(json.usage).toHaveProperty("completion_tokens");
+		expect(json.usage).toHaveProperty("total_tokens");
+
+		// Verify types are numbers
+		expect(typeof json.usage.prompt_tokens).toBe("number");
+		expect(typeof json.usage.completion_tokens).toBe("number");
+		expect(typeof json.usage.total_tokens).toBe("number");
+
+		// Most importantly: prompt_tokens should never be 0, even if provider returns 0
+		expect(json.usage.prompt_tokens).toBeGreaterThan(0);
+
+		// Completion tokens can be non-zero as set by mock
+		expect(json.usage.completion_tokens).toBeGreaterThan(0);
+
+		// Total tokens should be at least as large as prompt tokens
+		expect(json.usage.total_tokens).toBeGreaterThanOrEqual(
+			json.usage.prompt_tokens,
+		);
+	});
+
+	test("Prompt tokens are calculated for streaming when provider returns 0", async () => {
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: "openai/gpt-4o-mini",
+				messages: [
+					{
+						role: "user",
+						content: "ZERO_TOKENS streaming test message",
+					},
+				],
+				stream: true,
+			}),
+		});
+
+		expect(res.status).toBe(200);
+
+		const result = await readAll(res.body);
+
+		// Find a usage chunk
+		const usageChunk = result.chunks.find((chunk: any) => chunk.usage);
+		expect(usageChunk).toBeDefined();
+
+		if (usageChunk) {
+			// Verify prompt tokens are calculated and greater than 0
+			expect(usageChunk.usage.prompt_tokens).toBeGreaterThan(0);
+			expect(typeof usageChunk.usage.prompt_tokens).toBe("number");
+		}
+	});
+
+	test("GPT-5-nano responses API parameter handling", async () => {
+		const envVarName = getProviderEnvVar("openai");
+		const envVarValue = envVarName ? process.env[envVarName] : undefined;
+		if (!envVarValue) {
+			console.log(
+				"Skipping GPT-5-nano responses API test - no OpenAI API key provided",
+			);
+			return;
+		}
+
+		const requestId = generateTestRequestId();
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				"x-request-id": requestId,
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: "openai/gpt-5-nano",
+				messages: [
+					{
+						role: "user",
+						content: "What is 2+2? Think step by step.",
+					},
+				],
+				max_tokens: 100,
+				reasoning_effort: "medium",
+			}),
+		});
+
+		const json = await res.json();
+		if (logMode) {
+			console.log("GPT-5-nano response:", JSON.stringify(json, null, 2));
+		}
+
+		// Should succeed - no unsupported parameter error
+		expect(res.status).toBe(200);
+		validateResponse(json);
+
+		const log = await validateLogByRequestId(requestId);
+		expect(log.streamed).toBe(false);
+		expect(log.usedModel).toBe("openai/gpt-5-nano");
+		expect(log.usedModelMapping).toBe("gpt-5-nano");
+		expect(log.usedProvider).toBe("openai");
+
+		// Verify it's a reasoning model response
+		expect(json).toHaveProperty("usage");
+		if (json.usage.reasoning_tokens !== undefined) {
+			expect(typeof json.usage.reasoning_tokens).toBe("number");
+			expect(json.usage.reasoning_tokens).toBeGreaterThanOrEqual(0);
+		}
+
+		// Check for content - handle both string and object formats
+		expect(json.choices[0].message).toHaveProperty("content");
+	});
+
+	test("Success when requesting multi-provider model without prefix", async () => {
+		const multiProviderModel = models.find((m) => m.providers.length > 1);
+		if (!multiProviderModel) {
+			console.log(
+				"Skipping multi-provider test - no multi-provider models found",
+			);
+			return;
+		}
+
+		const requestId = generateTestRequestId();
+		const res = await app.request("/v1/chat/completions", {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+				"x-request-id": requestId,
+				Authorization: `Bearer real-token`,
+			},
+			body: JSON.stringify({
+				model: multiProviderModel.id,
+				messages: [
+					{
+						role: "user",
+						content: "Hello",
+					},
+				],
+			}),
+		});
+
+		expect(res.status).toBe(200);
+		const json = await res.json();
+		validateResponse(json);
+
+		const log = await validateLogByRequestId(requestId);
+		expect(log.streamed).toBe(false);
+	});
+});
+
+async function readAll(stream: ReadableStream<Uint8Array> | null): Promise<{
+	fullContent?: string;
+	hasContent: boolean;
+	eventCount: number;
+	hasValidSSE: boolean;
+	hasOpenAIFormat: boolean;
+	chunks: any[];
+	hasUsage: boolean;
+}> {
+	if (!stream) {
+		return {
+			hasContent: false,
+			eventCount: 0,
+			hasValidSSE: false,
+			hasOpenAIFormat: false,
+			chunks: [],
+			hasUsage: false,
+		};
+	}
+
+	const reader = stream.getReader();
+	let fullContent = "";
+	let eventCount = 0;
+	let hasValidSSE = false;
+	let hasContent = false;
+	let hasOpenAIFormat = true; // Assume true until proven otherwise
+	let hasUsage = false;
+	const chunks: any[] = [];
+
+	try {
+		while (true) {
+			const { done, value } = await reader.read();
+			if (done) {
+				break;
+			}
+
+			const chunk = new TextDecoder().decode(value);
+			fullContent += chunk;
+
+			const lines = chunk.split("\n");
+			for (const line of lines) {
+				if (line.startsWith("data: ")) {
+					eventCount++;
+					hasValidSSE = true;
+
+					if (line === "data: [DONE]") {
+						continue;
+					}
+
+					try {
+						const data = JSON.parse(line.substring(6));
+						chunks.push(data);
+
+						// Check if this chunk has OpenAI format
+						if (
+							!data.id ||
+							!data.object ||
+							data.object !== "chat.completion.chunk"
+						) {
+							hasOpenAIFormat = false;
+						}
+
+						// Check for content in OpenAI format (should be the primary format after transformation)
+						if (
+							data.choices?.[0]?.delta?.content ||
+							data.choices?.[0]?.finish_reason
+						) {
+							hasContent = true;
+						}
+
+						// Check for usage information
+						if (
+							data.usage &&
+							(data.usage.prompt_tokens !== null ||
+								data.usage.completion_tokens !== null ||
+								data.usage.total_tokens !== null)
+						) {
+							hasUsage = true;
+						}
+					} catch {}
+				}
+			}
+		}
+	} finally {
+		reader.releaseLock();
+	}
+
+	return {
+		fullContent,
+		hasContent,
+		eventCount,
+		hasValidSSE,
+		hasOpenAIFormat,
+		chunks,
+		hasUsage,
+	};
+}
